Inline tun::linux module into Linux TUN device manager
- Move tun_recv, parse_vnet_packet, segment_packet to linux.rs
- Move VirtioNetHdr struct and GSO constants
- Move VIRTIO_NET_HDR_SIZE and MAX_GSO_BUFFER_SIZE constants
- Remove tun::linux module entirely
- All Linux-specific TUN code now co-located in device manager

Benefits:
- Better code organization and separation of concerns
- tun crate focuses on trait and cross-platform types
- Platform-specific implementations live with device managers
